refactor: Migrate messaging system to EmbeddedMessage pool-based architecture
- Replace dynamic Message* with static EmbeddedMessage_t pool
- Update all FreeRTOS queues to use void* for embedded compatibility
- Implement MessagePool_Init/Allocate/Free for zero-allocation messaging
- Migrate all Message creation to EmbeddedMessage_Create/CreateWithPayload
- Update fsmTask, dispatcherTask, sensorAcqTask to use embedded messaging
- Remove all 'new Message()' and 'delete msg' calls
- Add RTOS_LOG_* macros for thread-safe debugging

Benefits:
- Zero dynamic allocation (heap fragmentation free)
- Predictable memory usage
- Thread-safe message passing
- MISRA-C compliant (no malloc/free in embedded critical paths)

+++ CM4/Core/Inc/cow.h
@@ -23,9 +23,9 @@ struct Position {
 };
 
 struct Acceleration {
-    float ax;
-    float ay;
-    float az;
+    double ax;
+    double ay;
+    double az;
 };
 
 class Cow {
